+++ include/metashell/shell.hpp
@@ -127,6 +127,9 @@ namespace metashell
 
     void show_cpp_errors(bool enabled_);
     bool show_cpp_errors() const;
+
+    void evaluate_metaprograms(bool enabled_);
+    bool evaluate_metaprograms() const;
   private:
     std::string _internal_dir;
     std::string _env_filename;
@@ -141,6 +144,7 @@ namespace metashell
     std::unique_ptr<iface::engine> _engine;
     bool _echo = false;
     bool _show_cpp_errors = true;
+    bool _evaluate_metaprograms = true;
 
     void init(command_processor_queue* cpq_);
     void rebuild_environment(const std::string& content_);